Adjust order of operations and target LMS_PROJECTS_DIR for all of them

+++ setup-image-build.sh
@@ -19,10 +19,26 @@ if ! groups | grep "\bdocker\b" >/dev/null; then
     exit 1
 fi
 
-# Clean the directory before cloning the repository
-rm -rf koha-testing-docker
+# Set up required environment variables
+export LMS_PROJECTS_DIR=~/.local/src/lmsc
+export LMS_PROJECTS_DIR="$LMS_PROJECTS_DIR"
+export LMS_SYNC_REPO=$LMS_PROJECTS_DIR/Koha-LMSCloud
+export LMS_KTD_HOME=$LMS_PROJECTS_DIR/LMSTestingDocker
+
+# Check if required environment variables are set up
+if [ -z "$LMS_PROJECTS_DIR" ] || [ -z "$LMS_SYNC_REPO" ] || [ -z "$LMS_KTD_HOME" ]; then
+    echo "Error: Required environment variables not found. Please set up the environment variables first." >&2
+    exit 1
+fi
+
+# Create the LMS_PROJECTS_DIR if it doesn't exist
+if [ ! -d "$LMS_PROJECTS_DIR" ]; then
+    mkdir -p $LMS_PROJECTS_DIR
+fi
 
 # Clone koha-testing-docker repository from GitLab
+cd $LMS_PROJECTS_DIR
+rm -rf koha-testing-docker
 git clone https://gitlab.com/koha-community/koha-testing-docker.git
 
 # Check if the repository was cloned successfully
@@ -38,21 +54,14 @@ if ! [ -x "$(command -v rsync)" ]; then
     apt-get install -y rsync
 fi
 
-# Set up required environment variables
-export LMS_PROJECTS_DIR=~/.local/src/lmsc
-export LMS_PROJECTS_DIR="$LMS_PROJECTS_DIR"
-export LMS_SYNC_REPO=$LMS_PROJECTS_DIR/Koha-LMSCloud
-export LMS_KTD_HOME=$LMS_PROJECTS_DIR/LMSTestingDocker
-
-# Check if required environment variables are set up
-if [ -z "$LMS_PROJECTS_DIR" ] || [ -z "$LMS_SYNC_REPO" ] || [ -z "$LMS_KTD_HOME" ]; then
-    echo "Error: Required environment variables not found. Please set up the environment variables first." >&2
-    exit 1
-fi
-
 # Set up LMSTestingDocker repository from GitHub
+cd $LMS_PROJECTS_DIR
+rm -rf LMSTestingDocker
 git clone https://github.com/LMSCloud/LMSTestingDocker.git $LMS_KTD_HOME
 
 cd $KTD_HOME
-git checkout -b k
+git checkout -b ktd-lms
+rsync -a --exclude='*.md' $LMS_KTD_HOME/* $KTD_HOME
+
+echo "LMSTestingDocker repository set up successfully."
